Reverting commit 6910d30eb01d66e4b844bed37fd1545cff17f15d.
----
An additional 4 commit messages were squashed into this commit:
--
Reverting commit 48cf45fcc11ac5078ed663827f936374808da0b8.
--
Reverting commit c1699b0aaa41c0602d4ae8573f13c4a2e7b2187e.
--
Reverting commit d395884e098c1d972cef3941fd8530b74b337796.
--
Reverting to commit d395884e098c1d972cef3941fd8530b74b337796.
--

Old repository SHA1: da84035144d96088d196cf885a5dc367f728d822

+++ source/src/protocols/fldsgn/filters/SecondaryStructureFilter.cc
@@ -182,7 +182,7 @@ SecondaryStructureFilter::set_blueprint( std::string const blueprint_file )
 	for ( core::Size strand=1; strand<=pairs.size(); ++strand ) {
 		core::Size const s1( pairs[strand]->s1() );
 		core::Size const s2( pairs[strand]->s2() );
-		int const shift( (int)(pairs[strand]->rgstr_shift()) );
+		int const shift( pairs[strand]->rgstr_shift() );
 		core::Size const len1( pairs[strand]->size1() );
 		core::Size const len2( pairs[strand]->size2() );
 		for ( core::Size res=1; res<=len1; ++res ) {
